pyproject.toml
[project]
name = "hello-django-ninja"
version = "0.1.0"
description = "A Django project with RBAC system based on DDD and django-ninja-extra"
readme = "README.md"
authors = [
    {name = "Taoweidong", email = "taowd@outlook.com"},
]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.3",
    "django-ninja-extra>=0.30.1",
    "py>=1.11.0",
    "pydantic==2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/nineaiyu/hello-django-ninja"
Repository = "https://github.com/nineaiyu/hello-django-ninja"
Documentation = "https://github.com/nineaiyu/hello-django-ninja#readme"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]